legacy/scaffold/pyproject.toml
[project]
name = "refinedc-copilot-scaffold"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Quinn Dougherty", email = "quinnd@riseup.net" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "jinja2>=3.1.5",
    "libclang>=18.1.1",
    "pydantic-ai[logfire]>=0.0.24",
    "python-dotenv>=1.0.1",
    "typer>=0.15.1",
]

[project.scripts]
test-pipeline = "refinedc_copilot_scaffold:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src/refinedc_copilot_scaffold"]

[dependency-groups]
dev = [
    "hypothesis>=6.126.0",
    "pyright>=1.1.394",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
]

# Add pytest configuration here
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
